Add error messages when we can't write to /usr/local/evm
Check for existance and writability of the directory, and give helpful
error messages if the checks fail.

Even though the readme says to use `chown` on the directory I think the
correct thing to check is writablity rather than ownership (since
advanced users could have the directory world-writable, or other strange
things, right?).