Assert that configured "path" is "/tmp" for Travis CI recipes
Emacs recipes for Travis CI require "path" to be configured to "/tmp",
otherwise complaining for example with:

| Warning: arch-dependent data dir (/tmp/emacs-24.2-travis/libexec/emacs/24.2/x86_64-unknown-linux-gnu/) does not exist.
| Warning: arch-independent data dir (/tmp/emacs-24.2-travis/share/emacs/24.2/etc/) does not exist.
| Warning: Lisp directory `/tmp/emacs-24.2-travis/share/emacs/24.2/lisp' does not exist.
| Warning: Lisp directory `/tmp/emacs-24.2-travis/share/emacs/24.2/leim' does not exist.
| Error: charsets directory not found:
| /tmp/emacs-24.2-travis/share/emacs/24.2/etc/charsets
| Emacs will not function correctly without the character map files.
| Please check your installation!

This was not apparent in EVM's current Travis CI setup because the
code path is not triggered by "emacs --version".  In addition, the
current Travis CI setup did not actually test that the Emacs binary
running was the version requested by $EVM_EMACS, instead relying on
manual inspection of the test logs.

This change adds assertions that "path" is configured to "/tmp" for
the Emacs recipes that require so and that the running Emacs binary is
the version requested by $EVM_EMACS, with the latter also serving as a
rudimentary functional test.