Make evm work when invoked through symlink
Previously if evm was invoked through a symlink instead of directly, __FILE__
pointed to the wrong file. It pointed to the symlink, not the actual evm script.
Thus the lib directory was wrong and 'require' failed.

This commit wraps __FILE__ in a call to realpath instead to resolve all possible
symlinks. See this Stackoverflow answer for more information:
https://stackoverflow.com/a/29494540/963029.